functions/list_blobs/requirements.txt
google-cloud-storage==2.16.0
functions-framework==3.7.0